requirements.txt
asgiref @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/asgiref_1729057877436/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961374486/work
certifi @ file:///C:/Users/task_176765924055489/croot/certifi_1767659363027/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792955/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744975868/work
Django @ file:///C:/b/abs_732pccxzov/croot/django_1745263601709/work
django-debug-toolbar==6.1.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1754514405054/work
idna @ file:///C:/miniconda3/conda-bld/idna_1761912043388/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030965/work
python-dotenv==1.2.1
requests @ file:///C:/miniconda3/conda-bld/requests_1762359611326/work
setuptools==78.1.1
sqlparse @ file:///C:/Users/dev-admin/buildout/croot/sqlparse_1739565583021/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
urllib3 @ file:///C:/miniconda3/conda-bld/urllib3_1765399243000/work
waitress @ file:///C:/Users/dev-admin/buildout/croot/waitress_1735864045188/work
wheel==0.45.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278300/work
